+++ src/osc/Bundler/CablesBundler.hpp
@@ -4,7 +4,7 @@
 
 struct CablesBundler : Bundler {
   CablesBundler();
-  CablesBundler(int64_t cableId, int32_t returnId);
+  CablesBundler(int64_t cableId, int64_t returnId);
 
-  void bundleCable(int64_t cableId, int32_t returnId = -1);
+  void bundleCable(int64_t cableId, int64_t returnId = -1);
 };
